More removal of open-watcom warnings
Mostly warnings removed by fixing prototypes.

More simplification to arch/i86/kernel/irqtab.c and
arch/i86/sibo/irqtab.c. More changes in arch/i86/sibo/irqtab.c
to be in sync with arch/i86/kernel/irqtab.c.

Code size was reduced by 32 bytes, data size increased 4 bytes
and bss size reduced by 4 bytes.

+++ elks/arch/i86/drivers/char/lp.c
@@ -162,7 +162,7 @@ int lp_char_polled(char c, unsigned int target)
     return 1;
 }
 
-int lp_write(struct inode *inode, struct file *file, char *buf, int count)
+size_t lp_write(struct inode *inode, struct file *file, char *buf, int count)
 {
     register char *chrsp;
 
@@ -180,7 +180,7 @@ int lp_write(struct inode *inode, struct file *file, char *buf, int count)
 	    break;
 	chrsp++;
     }
-    return (int) chrsp;
+    return (size_t)chrsp;
 }
 
 int lp_open(struct inode *inode, struct file *file)
